refactore: GenericJackson을 사용하여 타입 안정성 확보
